requirements.txt
streamlit==1.29.0
openai==0.28.1
chromadb==0.4.18
sentence-transformers==2.2.2
notion-client==2.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
discord.py==2.3.2
python-dateutil==2.8.2
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pptx==0.6.21
matplotlib==3.8.2
plotly==5.17.0
altair==5.2.0
